requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
django-filter>=24.2,<25.0
